Split out sim accuracy group, remove debug group
