+++ TLM/TLM/State/Options.cs
@@ -43,7 +43,7 @@ namespace TrafficManager.State {
         /// </remarks>
         public static bool Available = false;
 
-        public static bool instantEffects = true;
+        public static bool instantEffects;
         public static bool individualDrivingStyle = true;
         public static int recklessDrivers = 3;
