+++ TLM/TLM/TLM.csproj
@@ -152,7 +152,7 @@
     <Compile Include="State\OptionsTabs\GeneralTab_LocalisationGroup.cs" />
     <Compile Include="State\OptionsTabs\GameplayTab_VehicleBehaviourGroup.cs" />
     <Compile Include="State\OptionsTabs\GeneralTab_InterfaceGroup.cs" />
-    <Compile Include="State\OptionsTabs\GeneralTab_DebugGroup.cs" />
+    <Compile Include="State\OptionsTabs\GeneralTab_SimulationGroup.cs" />
     <Compile Include="State\OptionsTabs\MaintenanceTab_DespawnGroup.cs" />
     <Compile Include="State\VersionInfo.cs" />
     <Compile Include="UI\Helpers\ActionButton.cs" />
